+++ cpp/expression_cos.h
@@ -31,7 +31,7 @@ public:
     argument = &_argument;
     }
   
-  Number &count(int digits) const {
+  Number &count(long digits) const {
 
 #ifdef DEBUG
     LOG("COS");
@@ -48,42 +48,6 @@ public:
         getPi2(precision),
         getPi4(precision)
         );
-
-//  Number &numberSquare = argument->count(precision_argument);
-//  numberSquare.sqr();                 // x^{2}
-
-//  Number &number = *(new Number(1, precision));        
-//                                      // sum
-
-//  Number toAdd(1, precision);         // \frac{x^n}{n!}
-
-//  Number remainder(1, precision);
-//  remainder >>= precision;            // 2^{-a}
-
-//  int i = 2;
-//  while (toAdd.cmpAbs(remainder) > 0) {
-//    // TODO: Add check for cases with possibly wrong precision.
-//    
-//    //std::cout << "  to add 1: " << toAdd.asString() << std::endl;
-//    toAdd *= numberSquare;            // \frac{x^n}{(n-2)!}
-//    //std::cout << "  to add 2: " << toAdd.asString() << std::endl;
-//    toAdd /= Number(i*(i-1), precision);
-//                                      // \frac{x^n}{n!}
-//    //std::cout << "  to add 3: " << toAdd.asString() << std::endl;
-//    i += 2;
-//    if (!(i % 4)) {
-//      //std::cout << "to add: -" << toAdd.asString() << std::endl;
-//      number -= toAdd;
-//      }
-//    else {
-//      //std::cout << "to add: +" << toAdd.asString() << std::endl;
-//      number += toAdd;
-//      //std::cout << "to add: +" << toAdd.asString() << std::endl;
-//      }
-//    //std::cout << "number: " << number.asString() << std::endl;
-//    }
-//  
-//  return number;
     }
 
   Expression &clone() const {
